Numerics Rework

+++ FontoXPathCSharp/Expressions/Functions/BuiltInFunctionsFunctions.cs
@@ -12,19 +12,19 @@ public static class BuiltInFunctionsFunctions<TNode> where TNode : notnull
         return ISequence.ZipSingleton(sequences, nameArityTuple =>
         {
             var name = nameArityTuple[0].GetAs<QNameValue>();
-            var arity = nameArityTuple[1].GetAs<IntValue>();
+            var arity = nameArityTuple[1].GetAs<IntegerValue>();
 
             var functionProperties = staticContext?.LookupFunction(
                 name.Value.NamespaceUri,
                 name.Value.LocalName,
-                arity.Value
+                (int)arity.Value
             );
 
             if (functionProperties == null) return SequenceFactory.CreateEmpty();
 
             var functionItem = new FunctionValue<ISequence, TNode>(
                 functionProperties.ArgumentTypes,
-                arity.Value,
+                (int)arity.Value,
                 name.Value.LocalName,
                 name.Value.NamespaceUri!,
                 functionProperties.ReturnType,
